Finish up using LocationScouts instead of slot data
Sending items via cheat command now receives items local to Psychonauts
first and, only if there are no local copies of that item remaining does
it receive the items as if they are from another world.

Data specific to the current server connection is now cleaned up when
disconnecting. Ensuring a clean state when reconnecting or connecting to
a different server.

PsychoSeed.gen_psy_ids() is now more generic so that it can also be
called using LocationInfo scouted by the client. Additionally,
gen_psy_ids() now pre-sorts the input by location to ensure that the
same result is produced even if the locations are provided in a
different order (since it's not clear if there are guarantees to the
order of the LocationInfo in response to a LocationScouts request).